Harden updates and improve traffic monitoring

+++ NetSpeedMonitor/NetTrafficStatOC.h
@@ -2,16 +2,14 @@
 
 #import <Foundation/Foundation.h>
 
-#import "NetTrafficStatCpp.hpp"
-
 NS_ASSUME_NONNULL_BEGIN
 
 @interface NetTrafficStatOC : NSObject
 @property (nonatomic, strong) NSNumber *delta_ts_sec;
-@property (nonatomic, assign) NSInteger delta_ibytes;
-@property (nonatomic, assign) NSInteger delta_obytes;
-@property (nonatomic, assign) long long total_ibytes;
-@property (nonatomic, assign) long long total_obytes;
+@property (nonatomic, assign) unsigned long long delta_ibytes;
+@property (nonatomic, assign) unsigned long long delta_obytes;
+@property (nonatomic, assign) unsigned long long total_ibytes;
+@property (nonatomic, assign) unsigned long long total_obytes;
 @property (nonatomic, strong) NSNumber *ibytes_per_sec;
 @property (nonatomic, strong) NSNumber *obytes_per_sec;
 @property (nonatomic, assign) BOOL isUp;
